backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.51.0
python-dotenv==1.0.0
pydantic==2.4.2
aiofiles==23.2.1
httpx==0.27.0
supabase==2.8.1 